VTKU-51 : Extract common code
No functional changes.
