+++ sijoittelu-service/src/test/java/fi/vm/sade/sijoittelu/SijoitteluTilatJaKuvauksetTest.java
@@ -72,7 +72,7 @@ public class SijoitteluTilatJaKuvauksetTest {
             valintaTulosServiceResource,
             valintarekisteriService,
             new SijoitteluConfiguration(),
-            new SijoitteluajoResourcesLoader(tarjontaIntegrationService));
+            new SijoitteluajoResourcesLoader(tarjontaIntegrationService, valintarekisteriService));
 
         when(tarjontaIntegrationService.getHaunParametrit(hakuOid)).thenReturn(haunParametrit());
         when(tarjontaIntegrationService.getHakuByHakuOid(hakuOid)).thenReturn(tarjontaHaku());
